Create the SqlQueryGlobalSetting middleware
